Introduce dependency injection. We use StrongInject for "compile-time" DI.
Refactor various view models so that we eliminate parent coupling for the purpose of view model communication. Instead we use the recommended messenger pattern to communication application level events like [open resource], [connect to site], etc.

Introduce a new AppServices which is the service layer that our view models use to perform various tasks and use the WeakReferenceMessenger to relay events back to view models interested in the given events.