Introduce a new IViewModelFactory as a single one-stop shop for creating new view models.
I was getting tired of having to add Func<> to ctor parameters for specific view models, so in the interest of reducing ctor parameter count and to simplify design-time view model setup, it was decided to route most view model creation through this new interface.

A stub implementation is included for design-time purposes.

Also added inital implementation of sub-folder listing. Need to wire up node expansion to call this new service.