Implement on-demand loading of child nodes from a folder node
The TreeView is proving most un-cooperative and the bare documentation around this control is not helping. The expansion chevron does not actually show unless the folder node has child items, but I want a TreeView where the folder nodes load child nodes *on demand*.

As a workaround for that, I've gone with having an open action button alongside every node, that is only visible if the tree view item is being hovered over. Unfortunately, this too has issues where the button visibility of parent folder nodes of the currently hovered node will also be visible because I just cannot get the style selector to only pick out the *hover highlighted* control and ignore all the parent folder nodes that also happen to have the :pointerover pseudo-class.

This visual bug aside, this seems more workable as a way to perform an action on a node. So I've also ripped out the current (flaky) double-tap handlers and replaced it with a similar open button that's beside any resource node, which also has the same parented node button visibility problem.

*sigh*